Add analyze info and memory limit for operator
