+++ pkg/sql/colexec/merge/merge_test.go
@@ -96,7 +96,7 @@ func TestMerge(t *testing.T) {
 				}
 			}
 		}
-		require.Equal(t, mheap.Size(tc.proc.Mp), int64(0))
+		require.Equal(t, int64(0), mheap.Size(tc.proc.Mp))
 	}
 }
